mimic module-structure for tests
